% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls2model.PEPRMT.R
\name{hls2model.PEPRMT}
\alias{hls2model.PEPRMT}
\title{hls2model.PEPRMT}
\usage{
hls2model.PEPRMT(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  lat,
  lon,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date, end_date}{the start and end dates of the data to be downloaded (will only use the year part of the date)}

\item{lat, lon}{latitude and longitude in degrees}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{...}{additional arguments, currently ignored}
}
\description{
hls2model for PEPRMT
}
\author{
Abigail Lewis (add names)
}
