% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.PEPRMT.R
\name{model2netcdf.PEPRMT}
\alias{model2netcdf.PEPRMT}
\title{Function to convert PEPRMT model output to standard netCDF format}
\usage{
model2netcdf.PEPRMT(
  outdir,
  sitelat,
  sitelon,
  start_date,
  end_date,
  delete_raw = FALSE
)
}
\arguments{
\item{outdir}{Location of PEPRMT model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}

\item{delete_raw}{logical: remove out.csv after converting?}
}
\description{
Convert PEPRMT output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
Abigail Lewis
}
