% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.PEPRMT-package.R
\docType{package}
\name{PEcAn.PEPRMT-package}
\alias{PEcAn.PEPRMT}
\alias{PEcAn.PEPRMT-package}
\title{PEcAn.PEPRMT-package}
\description{
PEcAn support for the tidal biochemistry model PEPRMT
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pecanproject.github.io}
  \item Report bugs at \url{https://github.com/PecanProject/pecan/issues}
}

}
\author{
\strong{Maintainer}: Abigail Lewis \email{lewisa4@si.edu}

Authors:
\itemize{
  \item Patty Oikawa
  \item James Holmquist
}

}
\keyword{internal}
