% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.peprmt.R
\name{write.config.PEPRMT}
\alias{write.config.PEPRMT}
\title{Write PEPRMT configuration files}
\usage{
write.config.PEPRMT(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for PEPRMT for given run
}
\description{
Writes a PEPRMT config file.
}
\details{
Requires a pft xml object, a list of trait values for a single PEPRMT run,
and the name of the file to create
}
\author{
Rob Kooper, edited by Abby Lewis
}
